+++ discovery-provider/src/queries/search_es.py
@@ -445,163 +445,3 @@ def transform_tracks(tracks, users_by_id, current_user):
         tracks_out.append(track)
 
     return tracks_out
-
-
-# testing 1 2
-
-
-def _print_test_search(args):
-    print("\n\n==========", args)
-    found = search_es_full(args)
-
-    def print_tracks(title, tracks):
-        print(f"\n[ {title} ]")
-        for track in tracks:
-            print(
-                "   ",
-                [
-                    track["user"]["handle"],
-                    track["user"]["name"],
-                    track["title"],
-                    f"{track['repost_count']} reposts",
-                    f"{track['user']['follower_count']} followers",
-                    f"{track.get('_score')} score",
-                ],
-            )
-
-    def print_users(title, users):
-        print(f"\n[ {title} ]")
-        for user in users:
-            print(
-                "   ",
-                [
-                    user["handle"],
-                    user["name"],
-                    f"{user.get('follower_count')} followers",
-                    f"{user.get('is_verified')} verified",
-                    f"{user.get('_score')} score",
-                ],
-            )
-
-    print_tracks("tracks", found["tracks"])
-    # print_tracks("saved tracks", found["saved_tracks"])
-    print_users("users", found["users"])
-    # print_users("followed_users", found["followed_users"])
-
-
-if __name__ == "__main__":
-    # PYTHONPATH=. python3.9 src/queries/search_es.py
-
-    _print_test_search(
-        {
-            "query": "isaac pho",
-            "limit": 4,
-            "is_auto_complete": True,
-        }
-    )
-    _print_test_search(
-        {
-            "query": "isaac photo",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": False,
-        }
-    )
-    _print_test_search(
-        {
-            "query": "RAC wat",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": True,
-        }
-    )
-    _print_test_search(
-        {
-            "query": "RAC water",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": False,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "deadmau",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": False,
-        }
-    )
-
-    # should have disclosure at the top
-    _print_test_search(
-        {
-            "query": "waterfal",
-            "limit": 10,
-            "current_user_id": 1,
-            "is_auto_complete": True,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "closer 2 u ray",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": True,
-        }
-    )
-    _print_test_search(
-        {
-            "query": "raymont",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": True,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "low",
-            "limit": 4,
-            "current_user_id": 14,
-            "is_auto_complete": True,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "stereosteve guitar",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": True,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "skrillex",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": True,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "camo",
-            "limit": 4,
-            "is_auto_complete": True,
-        }
-    )
-
-    _print_test_search(
-        {
-            "query": "deadmau5",
-            "limit": 4,
-            "current_user_id": 1,
-            "is_auto_complete": False,
-        }
-    )
-
-    print("\n\n")
